Plane: make stab pitch down curve available to RC tuning

+++ ArduPlane/Plane.h
@@ -988,7 +988,7 @@ private:
     bool verify_nav_script_time(const AP_Mission::Mission_Command& cmd);
 #endif
 
-    static float apply_pitch_curve(float v, int8_t curve);
+    static float apply_pitch_curve(float v, float curve);
 
     // commands.cpp
     void set_guided_WP(void);
